requirements.txt
pytz~=2023.3.post1
requests~=2.31.0
prettytable~=3.9.0
python-dotenv~=1.0.1